+++ .gitignore
@@ -1,3 +1,5 @@
+dist/
+
 # Logs
 logs
 *.log
